# app/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server We Will Guide You server.",
        "version": "1.0.0",
        "title": "Swagger We Will Guide You",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
...
    "schemes": [
        "https"
    ],
    "paths": {
...
        "/manage/websiteSetting/": {
            "get": {
                "tags": [
                    "WebsiteSetting"
                ],
                "summary": "Get WebsiteSetting",
                "description": "Returns a single WebsiteSetting model.",
                "operationId": "getWebsiteSetting",
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/DetailWebsiteSettingAdminApiModel"
                        }
                    }
                },
                "security": [{
                    "Bearer": []
                }]
            },
            "put": {
                "tags": [
                    "WebsiteSetting"
                ],
                "summary": "Update an existing WebsiteSetting",
                "description": "Update an existing WebsiteSetting",
                "operationId": "updateWebsiteSetting",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Modified WebsiteSetting model.",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/UpdateWebsiteSettingAdminApiModel"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "WebsiteSetting updated successfully",
                        "schema": {
                            "$ref": "#/definitions/DetailWebsiteSettingAdminApiModel"
                        }
                    }
                },
                "security": [{
                    "Bearer": []
                }]
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "in": "header",
            "name": "Authorization"
        }
    },
    "definitions": {
...
        "ResponseMessage": {
            "type": "object",
            "properties": {
                "eventId": {
                    "type": "number"
                },
                "messageId": {
                    "type": "number"
                },
                "message": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "error",
                        "warning",
                        "info"
                    ]
                }
            }
        },
...
        "UpdateWebsiteSettingAdminApiModel": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string"
                },
                "contactUsText": {
                    "type": "string"
                },
                "aboutUsText": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "keywords": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "DetailWebsiteSettingAdminApiModel": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "object",
                    "properties": {
                        "title": {
                            "type": "string"
                        },
                        "contactUsText": {
                            "type": "string"
                        },
                        "aboutUsText": {
                            "type": "string"
                        },
                        "description": {
                            "type": "string"
                        },
                        "keywords": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ResponseMessage"
                    }
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
...
}